backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.1.0
numpy==1.24.3
matplotlib==3.8.2
plotly==5.18.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2